requirements.txt
pandas
numpy
scikit-learn
matplotlib
tqdm
Pillow